+++ th2_data_services/provider/v5/commands/http.py
@@ -37,7 +37,7 @@ logger.setLevel(logging.DEBUG)
 class GetEventById(IHTTPProvider5Command, ProviderAdaptableCommand):
     """A Class-Command for request to rpt-data-provider.
 
-    It retrieves the event by id.
+    It retrieves the event by id with `attachedMessageIds` list.
 
     Returns:
         dict: Th2 event.
@@ -79,7 +79,7 @@ class GetEventById(IHTTPProvider5Command, ProviderAdaptableCommand):
 class GetEventsById(IHTTPProvider5Command, ProviderAdaptableCommand):
     """A Class-Command for request to rpt-data-provider.
 
-    It retrieves the events by ids.
+    It retrieves the events by ids with `attachedMessageIds` list.
 
     Returns:
         List[dict]: Th2 events.
@@ -323,7 +323,7 @@ class GetEvents(IHTTPProvider5Command, ProviderAdaptableCommand):
     def handle(self, data_source: HTTPProvider5DataSource) -> Data:  # noqa: D102
         source = partial(self.__handle_stream, data_source)
         adapter = SSEAdapter()
-        return Data(source, cache=self._cache).map(adapter.handle)
+        return Data(source).map(adapter.handle).use_cache(self._cache)
 
     def __handle_stream(self, data_source: HTTPProvider5DataSource) -> Generator[dict, None, None]:
         stream = GetEventsSSEEvents(
@@ -348,6 +348,9 @@ class GetMessageById(IHTTPProvider5Command, ProviderAdaptableCommand):
     """A Class-Command for request to rpt-data-provider.
 
     It retrieves the message by id.
+
+    Please note, Provider5 doesn't return `attachedEventIds`. It will be == [].
+    It's expected that Provider7 will be support it.
 
     Returns:
         dict: Th2 message.
@@ -391,6 +394,9 @@ class GetMessagesById(IHTTPProvider5Command, ProviderAdaptableCommand):
     """A Class-Command for request to rpt-data-provider.
 
     It retrieves the messages by ids.
+
+    Please note, Provider5 doesn't return `attachedEventIds`. It will be == [].
+    It's expected that Provider7 will be support it.
 
     Returns:
         List[dict]: Th2 messages.
@@ -665,7 +671,7 @@ class GetMessages(IHTTPProvider5Command, ProviderAdaptableCommand):
     def handle(self, data_source: HTTPProvider5DataSource) -> Data:  # noqa: D102
         source = partial(self.__handle_stream, data_source)
         adapter = SSEAdapter()
-        return Data(source, cache=self._cache).map(adapter.handle)
+        return Data(source).map(adapter.handle).use_cache(self._cache)
 
     def __handle_stream(self, data_source: HTTPProvider5DataSource) -> Generator[dict, None, None]:
         stream = GetMessagesSSEEvents(