+++ th2_data_services/provider/v5/events_tree/events_tree_collection.py
@@ -20,6 +20,7 @@ from th2_data_services.events_tree.exceptions import FieldIsNotExist
 from th2_data_services.provider.interfaces.struct import IEventStruct
 from th2_data_services.provider.v5.data_source import HTTPProvider5DataSource, GRPCProvider5DataSource
 from th2_data_services.provider.v5.struct import provider5_event_struct
+from th2_data_services.provider.v5.stub_builder import provider5_event_stub_builder
 
 
 class EventsTreeCollectionProvider5(EventsTreeCollection):
@@ -81,3 +82,14 @@ class EventsTreeCollectionProvider5(EventsTreeCollection):
             Parent event id.
         """
         return event.get(self._event_struct.PARENT_EVENT_ID)
+
+    def _build_stub_event(self, id_: str):
+        """Builds stub event.
+
+        Args:
+            id_: Event Id.
+
+        Returns:
+            Stub event.
+        """
+        return provider5_event_stub_builder.build({self._event_struct.EVENT_ID: id_})